Move alembic models from models/ to models/alembic
